Separate FutureProcessor handlers and helpers
