+++ packages/core/src/new-api/internal/new-execution/future-processor/handlers/query-static-call.ts
@@ -11,8 +11,7 @@ import {
   JournalMessageType,
 } from "../../types/messages";
 import { NetworkInteractionType } from "../../types/network-interaction";
-
-import { runStaticCall } from "./network-interaction-execution";
+import { runStaticCall } from "../helpers/network-interaction-execution";
 
 /**
  * Runs a static call and returns a message indicating its completion.
